[#29][restaurant] move: 패키지 위치 이동
