+++ restaurant/src/main/java/table/eat/now/restaurant/restaurant/domain/entity/ContactInfo.java
@@ -2,10 +2,9 @@
  * @author : jieun(je-pa)
  * @Date : 2025. 04. 08.
  */
-package table.eat.now.restaurant.domain.entity;
+package table.eat.now.restaurant.restaurant.domain.entity;
 
 import jakarta.persistence.Embeddable;
-import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -14,8 +13,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class OperatingTime {
-  private LocalDateTime openingAt;
-  private LocalDateTime closingAt;
+public class ContactInfo {
+  private String contactNumber;
+  private String address;
 }
