+++ restaurant/src/main/java/table/eat/now/restaurant/restaurant/domain/entity/OperatingTime.java
@@ -2,9 +2,10 @@
  * @author : jieun(je-pa)
  * @Date : 2025. 04. 08.
  */
-package table.eat.now.restaurant.domain.entity;
+package table.eat.now.restaurant.restaurant.domain.entity;
 
 import jakarta.persistence.Embeddable;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -13,8 +14,8 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
-public class ContactInfo {
-  private String contactNumber;
-  private String address;
+public class OperatingTime {
+  private LocalDateTime openingAt;
+  private LocalDateTime closingAt;
 }
